BOOTCAMP PROJECT
A game of a sweet pink bouncing ball.
Your task is simple- collect all of the coloured dots appearing on the screen while fighting a desperate and unfair battle against the physics. But watch out.
Dont step off to the spikes that appeared at the bottom or they will kill you. You don't want that.
To help yourself navigate the game using A, S, D and move one of the platforms with your mouse.
The odds are against you and the physics are a bit wonky so...
GOOD LUCK!
And remember- just like in real life you work hard to defy death and score some points with every bounce but for what?
What is the meaning of life?
Is 42 good enough for this little ball?